@@ -1,29 +1,27 @@
 {
-  "name": "Sell-Hive Website",
-  "description": "",
   "require": {
-    "hassankhan/config": "~2.0.0",
-    "monolog/monolog": "^1.0",
-    "php-di/php-di": "~6.0.5",
-    "php-di/slim-bridge": "~2.0.0",
+    "hassankhan/config": "~3.1.0",
+    "monolog/monolog": "~1.0",
+    "php-di/php-di": "~7.0.6",
+    "php-di/slim-bridge": "~3.4.0",
     "phpmailer/phpmailer": "~6.0",
-    "propel/propel": "~2.0@dev",
-    "respect/validation": "1.1.*",
-    "setasign/fpdf": "1.8.*",
-    "slim/slim": "~3.12.0",
-    "slim/twig-view": "~2.4.0",
-    "symfony/yaml": "~4.2.3",
-    "twig/twig": "~2.6.2",
+    "propel/propel": "~2.0@beta",
+    "respect/validation": "~2.2.4",
+    "setasign/fpdf": "~1.8.0",
+    "slim/slim": "~4.9.0",
+    "slim/psr7": "~1.6.1",
+    "slim/twig-view": "~3.3.0",
+    "symfony/yaml": "~v5.4.31",
+    "twig/twig": "~v3.8.0",
     "twitter/bootstrap": "~4.2.1",
-    "zircote/swagger-php": "^3.0",
-    "doctrine/annotations": "~1.8.0"
+    "zircote/swagger-php": "~3.0.0",
+    "doctrine/annotations": "~2.0.0"
   },
   "require-dev": {
-    "phpunit/phpunit": "^6"
+    "phpunit/phpunit": "~10"
   },
   "autoload": {
     "classmap": [
-      "SellHiveApp.php",
       "controller/",
       "propel/model/",
       "service/",
