@@ -1,4 +1,4 @@
-<?php
+<?php /** @noinspection PhpUnused */
 
 use Psr\Http\Message\ServerRequestInterface as Request;
 use Psr\Http\Message\ResponseInterface as Response;
@@ -8,6 +8,17 @@ use Noodlehaus\Config;
 
 class AuthController
 {
+    private Logger $logger;
+    private Config $config;
+    private MailService $mail;
+
+    public function __construct(Logger $logger, Config $config, MailService $mail)
+    {
+        $this->logger = $logger;
+        $this->config = $config;
+        $this->mail = $mail;
+    }
+
     /**
      * @OA\Get(
      *     path="/backend/auth",
@@ -22,18 +33,22 @@ class AuthController
      *     )
      * )
      */
-    public function isAuthenticated(Request $request, Response $response, Logger $logger)
+    public function isAuthenticated(Request $request, Response $response): Response
     {
-        $logger->debug('=== AuthController:isAuthenticated(...) ===');
+        $this->logger->debug('=== AuthController:isAuthenticated(...) ===');
 
         $out = array();
 
         $out['authenticated'] = isset($_SESSION['user']);
         $out['admin'] = isset($_SESSION['user']) && $_SESSION['user'] == -1;
 
-        $logger->debug('output', $out);
+        $this->logger->debug('output', $out);
 
-        return $response->withJson($out, 200, JSON_PRETTY_PRINT);
+        $response->getBody()->write(json_encode($out, JSON_PRETTY_PRINT));
+
+        return $response
+            ->withHeader('Content-Type', 'application/json')
+            ->withStatus(200);
     }
 
     /**
@@ -57,9 +72,9 @@ class AuthController
      *     @OA\Response(response="400", description="login failed")
      * )
      */
-    public function login(Request $request, Response $response, Logger $logger, Config $config)
+    public function login(Request $request, Response $response)
     {
-        $logger->debug('=== AuthController:login(...) ===');
+        $this->logger->debug('=== AuthController:login(...) ===');
 
         $out = array();
         $out['login'] = true;
@@ -67,53 +82,57 @@ class AuthController
 
         $in = $request->getParsedBody();
 
-        $logger->debug('input', isset($in) ? $in : array());
+        $this->logger->debug('input', $in ?? array());
 
         if (v::not(v::keySet(v::key('mail'), v::key('password')))->validate($in)) {
-            $logger->debug('missing input "mail" and/or "password"');
+            $this->logger->debug('missing input "mail" and/or "password"');
             return $response->withStatus(400);
         }
 
         if (v::equals('admin')->validate($in['mail'])) {
-            $logger->debug('is admin login');
-            $out['admin'] = $config->get('admin.active');
+            $this->logger->debug('is admin login');
+            $out['admin'] = $this->config->get('admin.active');
         } elseif (v::email()->length(1, 254)->validate($in['mail'])) {
-            $logger->debug('is user login');
+            $this->logger->debug('is user login');
             $out['mail'] = 'valid';
         } else {
-            $logger->debug('invalid mail');
+            $this->logger->debug('invalid mail');
             $out['login'] = false;
         }
 
         if (!v::alnum()->length(1, 64)->validate($in['password'])) {
-            $logger->debug('invalid password');
+            $this->logger->debug('invalid password');
             $out['login'] = false;
         }
 
         if ($out['login'] && $out['admin']) {
-            if (v::equals($config->get('admin.password'))->validate($in['password'])) {
+            if (v::equals($this->config->get('admin.password'))->validate($in['password'])) {
                 $_SESSION['user'] = -1;
-                $logger->debug('admin login success');
+                $this->logger->debug('admin login success');
             } else {
                 $out['login'] = false;
-                $logger->debug('admin login failed');
+                $this->logger->debug('admin login failed');
             }
         } elseif ($out['login']) {
             $seller = SellerQuery::create()->filterByMail($in['mail'])->filterByPassword($in['password'])->findOne();
             if ($seller == null) {
                 $out['login'] = false;
-                $logger->debug('user login failed');
+                $this->logger->debug('user login failed');
             } else {
                 $_SESSION['user'] = $seller->getId();
-                $logger->debug('user login success');
+                $this->logger->debug('user login success');
             }
         } else {
-            $logger->debug('login invalid');
+            $this->logger->debug('login invalid');
         }
 
-        $logger->debug('output', $out);
+        $this->logger->debug('output', $out);
 
-        return $response->withJson($out, 200, JSON_PRETTY_PRINT);
+        $response->getBody()->write(json_encode($out, JSON_PRETTY_PRINT));
+
+        return $response
+            ->withHeader('Content-Type', 'application/json')
+            ->withStatus(200);
     }
 
     /**
@@ -123,16 +142,16 @@ class AuthController
      *     @OA\Response(response="200", description="logged out successfully or logout not required")
      * )
      */
-    public function logout(Request $request, Response $response, Logger $logger)
+    public function logout(Request $request, Response $response)
     {
-        $logger->debug('=== AuthController:logout(...) ===');
+        $this->logger->debug('=== AuthController:logout(...) ===');
 
         if (isset($_SESSION['user'])) {
             session_destroy();
-            $logger->debug('logout success');
+            $this->logger->debug('logout success');
             return $response->withStatus(200);
         } else {
-            $logger->debug('logout not required');
+            $this->logger->debug('logout not required');
             return $response->withStatus(200);
         }
     }
@@ -155,9 +174,9 @@ class AuthController
      *     )
      * )
      */
-    public function remind(Request $request, Response $response, Logger $logger, Config $config, MailService $mail)
+    public function remind(Request $request, Response $response)
     {
-        $logger->debug('=== AuthController:remind(...) ===');
+        $this->logger->debug('=== AuthController:remind(...) ===');
 
         /* TODO : recaptcha prüfen
         $recaptcha = new \ReCaptcha\ReCaptcha($secret);
@@ -174,23 +193,27 @@ class AuthController
 
         $in = $request->getParsedBody();
 
-        $logger->debug('input', isset($in) ? $in : array());
+        $this->logger->debug('input', isset($in) ? $in : array());
 
         if (v::email()->length(1, 254)->validate($in['mail'])) {
             $seller = SellerQuery::create()->filterByMail($in['mail'])->findOne();
 
             if ($seller == null) {
-                $logger->debug('seller not found');
+                $this->logger->debug('seller not found');
             } else {
                 $seller->genPassword();
                 $seller->save();
 
-                $out['mailed'] = $mail->sendPasswordReminderToSeller($seller);
+                $out['mailed'] = $this->mail->sendPasswordReminderToSeller($seller);
             }
         }
 
-        $logger->debug('output', $out);
+        $this->logger->debug('output', $out);
 
-        return $response->withJson($out, 200, JSON_PRETTY_PRINT);
+        $response->getBody()->write(json_encode($out, JSON_PRETTY_PRINT));
+
+        return $response
+            ->withHeader('Content-Type', 'application/json')
+            ->withStatus(200);
     }
 }
