@@ -1,4 +1,4 @@
-<?php
+<?php /** @noinspection PhpUnused */
 
 use Psr\Http\Message\ServerRequestInterface as Request;
 use Psr\Http\Message\ResponseInterface as Response;
@@ -6,14 +6,28 @@ use Psr\Log\LoggerInterface as Logger;
 use Respect\Validation\Validator as v;
 use Propel\Runtime\ActiveQuery\Criteria as c;
 use Noodlehaus\Config;
+use Slim\Routing\RouteContext;
 
 class SellerLimitController
 {
-    public function get(Request $request, Response $response, Logger $logger, AuthService $auth, Config $config)
-    {
-        $logger->debug('=== SellerLimitController:get(...) ===');
+    private Logger $logger;
+    private Config $config;
+    private AuthService $auth;
+    private MailService $mail;
 
-        if ($auth->isNoUser()) {
+    public function __construct(Logger $logger, Config $config, AuthService $auth, MailService $mail)
+    {
+        $this->logger = $logger;
+        $this->config = $config;
+        $this->auth = $auth;
+        $this->mail = $mail;
+    }
+
+    public function get(Request $request, Response $response)
+    {
+        $this->logger->debug('=== SellerLimitController:get(...) ===');
+
+        if ($this->auth->isNoUser()) {
             return $response->withStatus(403);
         }
 
@@ -23,27 +37,31 @@ class SellerLimitController
         $out['current'] = $seller->getLimit();
         $out['requested'] = $seller->getLimitRequest();
 
-        return $response->withJson($out, 200, JSON_PRETTY_PRINT);
+        $response->getBody()->write(json_encode($out, JSON_PRETTY_PRINT));
+
+        return $response
+            ->withHeader('Content-Type', 'application/json')
+            ->withStatus(200);
     }
 
-    public function openRequest(Request $request, Response $response, Logger $logger, AuthService $auth, MailService $mail, Config $config)
+    public function openRequest(Request $request, Response $response)
     {
-        $logger->debug('=== SellerLimitController:openRequest(...) ===');
+        $this->logger->debug('=== SellerLimitController:openRequest(...) ===');
 
-        if ($auth->isNoUser()) {
+        if ($this->auth->isNoUser()) {
             return $response->withStatus(403);
         }
 
         $in = $request->getParsedBody();
 
-        $logger->debug('input', $in);
+        $this->logger->debug('input', $in);
 
         $out = array();
         $out['valid'] = true;
         $out['saved'] = false;
         $out['mailed'] = false;
 
-        $maxStep = $config->get('seller.limit.maxStep');
+        $maxStep = $this->config->get('seller.limit.maxStep');
         $seller = SellerQuery::create()->requireOneById($_SESSION['user']);
         $itemCount = $seller->countItems();
 
@@ -52,29 +70,33 @@ class SellerLimitController
         }
 
         if ($out['valid']) {
-            $logger->debug('open limit request');
+            $this->logger->debug('open limit request');
 
             $seller->setLimitRequest($in['limit']);
             $seller->save();
 
             $out['saved'] = true;
-            $logger->debug('seller saved');
+            $this->logger->debug('seller saved');
 
-            $out['mailed'] = $mail->sendLimitRequestToAdmin($seller);
+            $out['mailed'] = $this->mail->sendLimitRequestToAdmin($seller);
         } else {
-            $logger->debug('data invalid');
+            $this->logger->debug('data invalid');
         }
 
-        $logger->debug('output', $out);
+        $this->logger->debug('output', $out);
 
-        return $response->withJson($out, 200, JSON_PRETTY_PRINT);
+        $response->getBody()->write(json_encode($out, JSON_PRETTY_PRINT));
+
+        return $response
+            ->withHeader('Content-Type', 'application/json')
+            ->withStatus(200);
     }
 
-    public function resetLimits(Request $request, Response $response, Logger $logger, Config $config)
+    public function resetLimits(Request $request, Response $response)
     {
-        $logger->debug('=== SellerLimitController:resetLimits(...) ===');
+        $this->logger->debug('=== SellerLimitController:resetLimits(...) ===');
 
-        if ($this->secretIsInvalid($request, $config)) {
+        if ($this->secretIsInvalid($request, $this->config)) {
             return $response->withStatus(403);
         }
 
@@ -83,7 +105,7 @@ class SellerLimitController
         foreach ($sellers as $seller) {
             $sellerName = $seller->getName();
             $sellerId = $seller->getId();
-            $limitTill = $seller->getLimitTill() ;
+            $limitTill = $seller->getLimitTill();
 
             if ($limitTill < $today) {
                 $limit = $seller->getLimit();
@@ -92,23 +114,23 @@ class SellerLimitController
                 $seller->setLimitTill(null);
                 $seller->save();
 
-                $logger->info("Limit für $sellerName ($sellerId) von $limit auf $itemCount zurückgesetzt.");
+                $this->logger->info("Limit für $sellerName ($sellerId) von $limit auf $itemCount zurückgesetzt.");
             } else {
-                $logger->debug("Limit für $sellerName ($sellerId) bis " . date_format($limitTill, 'Y-m-d') . " beibehalten.");
+                $this->logger->debug("Limit für $sellerName ($sellerId) bis " . date_format($limitTill, 'Y-m-d') . " beibehalten.");
             }
         }
 
         return $response->withStatus(200);
     }
 
-    private function secretIsInvalid(Request $request, Config $config)
+    private function secretIsInvalid(Request $request): bool
     {
-        $validSecret = $config->get('seller.limit.secret');
-        $submittedSecret = $request->getAttribute('route')->getArgument('secret', 'null');
+        $validSecret = $this->config->get('seller.limit.secret');
+        $submittedSecret = RouteContext::fromRequest($request)->getRoute()->getArgument('secret', 'null');
         $valid = v::equals($validSecret)->validate($submittedSecret);
 
         if (!$valid) {
-            $logger->warn('invalid secret "' . $submittedSecret . '"');
+            $this->logger->warn('invalid secret "' . $submittedSecret . '"');
         }
 
         return !$valid;
