@@ -7,14 +7,25 @@ use Slim\Views\Twig as Twig;
 
 class ItemTableController
 {
-    public function show(Request $request, Response $response, Logger $logger, Twig $twig, ContextService $contextService)
-    {
-        $logger->debug('=== ItemTableController:show(...) ===');
+    private ContextService $contextService;
+    private Logger $logger;
+    private Twig $twig;
 
-        $context = $contextService->getGlobal();
+    public function __construct(ContextService $contextService, Logger $logger, Twig $twig)
+    {
+        $this->contextService = $contextService;
+        $this->logger = $logger;
+        $this->twig = $twig;
+    }
+
+    public function show(Request $request, Response $response): Response
+    {
+        $this->logger->debug('=== ItemTableController:show(...) ===');
+
+        $context = $this->contextService->getGlobal();
 
         $context['title'] = 'Spiele';
 
-        return $twig->render($response, 'content/itemTable.twig', $context);
+        return $this->twig->render($response, 'content/itemTable.twig', $context);
     }
 }
