@@ -7,11 +7,22 @@ use Slim\Views\Twig as Twig;
 
 class OpenLimitRequestModalController
 {
-    public function show(Request $request, Response $response, Logger $logger, Twig $twig, ContextService $contextService)
-    {
-        $logger->debug('=== OpenLimitRequestModalController:show(...) ===');
+    private ContextService $contextService;
+    private Logger $logger;
+    private Twig $twig;
 
-        $context = $contextService->getGlobal();
+    public function __construct(ContextService $contextService, Logger $logger, Twig $twig)
+    {
+        $this->contextService = $contextService;
+        $this->logger = $logger;
+        $this->twig = $twig;
+    }
+
+    public function show(Request $request, Response $response): Response
+    {
+        $this->logger->debug('=== OpenLimitRequestModalController:show(...) ===');
+
+        $context = $this->contextService->getGlobal();
 
         $context['title'] = 'Mehr Spiele?';
 
@@ -23,6 +34,6 @@ class OpenLimitRequestModalController
         $context['submit'] = 'abschicken';
         $context['cancel'] = 'abbrechen';
 
-        return $twig->render($response, 'modal/openLimitRequest.twig', $context);
+        return $this->twig->render($response, 'modal/openLimitRequest.twig', $context);
     }
 }
