@@ -7,11 +7,22 @@ use Slim\Views\Twig as Twig;
 
 class PrintSettingsModalController
 {
-    public function show(Request $request, Response $response, Logger $logger, Twig $twig, ContextService $contextService)
-    {
-        $logger->debug('=== PrintSettingsModalController:show(...) ===');
+    private ContextService $contextService;
+    private Logger $logger;
+    private Twig $twig;
 
-        $context = $contextService->getGlobal();
+    public function __construct(ContextService $contextService, Logger $logger, Twig $twig)
+    {
+        $this->contextService = $contextService;
+        $this->logger = $logger;
+        $this->twig = $twig;
+    }
+
+    public function show(Request $request, Response $response): Response
+    {
+        $this->logger->debug('=== PrintSettingsModalController:show(...) ===');
+
+        $context = $this->contextService->getGlobal();
 
         $context['title'] = 'Ausrichtung anpassen';
 
@@ -25,7 +36,7 @@ class PrintSettingsModalController
         $context['pageInit']['y']['invalid'] = 'Der Wert muss zwischen X mm und X mm liegen.';
         $context['pageInit']['y']['img'] = 'img/PrintSettingsPageInitY.png';
 
-        $context['labelSpace']['x']['label'] = 'Wagerechter Abstand der Etiketten';
+        $context['labelSpace']['x']['label'] = 'Waagerechter Abstand der Etiketten';
         $context['labelSpace']['x']['help'] = 'Wert zwischen X mm und X mm';
         $context['labelSpace']['x']['invalid'] = 'Der Wert muss zwischen X mm und X mm liegen.';
         $context['labelSpace']['x']['img'] = 'img/PrintSettingsLabelSpaceX.png';
@@ -42,6 +53,6 @@ class PrintSettingsModalController
         $context['submit'] = 'speichern';
         $context['cancel'] = 'abbrechen';
 
-        return $twig->render($response, 'modal/printSettings.twig', $context);
+        return $this->twig->render($response, 'modal/printSettings.twig', $context);
     }
 }
