@@ -3,17 +3,29 @@
 use Psr\Http\Message\ServerRequestInterface as Request;
 use Psr\Http\Message\ResponseInterface as Response;
 use Psr\Log\LoggerInterface as Logger;
+use Slim\Routing\RouteContext;
 use Slim\Views\Twig as Twig;
 
 class SellerEditorModalController
 {
-    public function show(Request $request, Response $response, Logger $logger, Twig $twig, ContextService $contextService)
+    private ContextService $contextService;
+    private Logger $logger;
+    private Twig $twig;
+
+    public function __construct(ContextService $contextService, Logger $logger, Twig $twig)
     {
-        $logger->debug('=== SellerEditorModalController:show(...) ===');
+        $this->contextService = $contextService;
+        $this->logger = $logger;
+        $this->twig = $twig;
+    }
+
+    public function show(Request $request, Response $response): Response
+    {
+        $this->logger->debug('=== SellerEditorModalController:show(...) ===');
 
         // TODO validate args
 
-        $context = $contextService->getGlobal();
+        $context = $this->contextService->getGlobal();
 
         $context['title'] = 'Verkäufer bearbeiten';
 
@@ -29,8 +41,8 @@ class SellerEditorModalController
         $context['submit'] = 'abschicken';
         $context['cancel'] = 'abbrechen';
 
-        $context['sellerId'] = $request->getAttribute('route')->getArgument('sellerId');
+        $context['sellerId'] = RouteContext::fromRequest($request)->getRoute()->getArgument('sellerId');
 
-        return $twig->render($response, 'modal/SellerEditor.twig', $context);
+        return $this->twig->render($response, 'modal/SellerEditor.twig', $context);
     }
 }
